+++ deboucled.user.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name        Déboucled
 // @namespace   deboucledjvcom
-// @version     2.66.0
+// @version     2.65.1
 // @downloadURL https://github.com/Rand0max/deboucled/raw/master/deboucled.user.js
 // @updateURL   https://github.com/Rand0max/deboucled/raw/master/deboucled.meta.js
 // @author      Rand0max
@@ -73,5 +73,5 @@
 // @require     profile.js
 // @require     extras.js
 // @require     main.js
-// @run-at      document-body
+// @run-at      document-end
 // ==/UserScript==